Replace Fusebox setClientMetadata handshake with ReactNativeApplicationModel
