+++ front_end/core/sdk/ReactNativeApplicationModel.ts
@@ -6,6 +6,7 @@
 import type * as ProtocolProxyApi from '../../generated/protocol-proxy-api.js';
 import type * as Protocol from '../../generated/protocol.js';
 
+import * as Host from '../../core/host/host.js';
 import {Capability, type Target} from './Target.js';
 import {SDKModel} from './SDKModel.js';
 
@@ -18,9 +19,14 @@ export class ReactNativeApplicationModel extends SDKModel<EventTypes> implements
   constructor(target: Target) {
     super(target);
 
+    Host.rnPerfMetrics.fuseboxSetClientMetadataStarted();
+
     this.#enabled = false;
     this.#agent = target.reactNativeApplicationAgent();
     target.registerReactNativeApplicationDispatcher(this);
+
+    // Auto-init. Paired with registering this model immediately in rn_fusebox.ts.
+    this.ensureEnabled();
   }
 
   ensureEnabled(): void {
@@ -28,7 +34,16 @@ export class ReactNativeApplicationModel extends SDKModel<EventTypes> implements
       return;
     }
 
-    void this.#agent.invoke_enable();
+    void this.#agent.invoke_enable()
+      .then(result => {
+        const maybeError = result.getError();
+        const success = !maybeError;
+        Host.rnPerfMetrics.fuseboxSetClientMetadataFinished(success, maybeError);
+      })
+      .catch(reason => {
+        const success = false;
+        Host.rnPerfMetrics.fuseboxSetClientMetadataFinished(success, reason);
+      });
     this.#enabled = true;
   }
 
@@ -45,11 +60,3 @@ export const enum Events {
 export type EventTypes = {
   [Events.MetadataUpdated]: Protocol.ReactNativeApplication.MetadataUpdatedEvent,
 };
-
-SDKModel.register(
-  ReactNativeApplicationModel,
-  {
-    capabilities: Capability.None,
-    autostart: true,
-  },
-);
